Lambda function to detect CloudTrail being disabled
Runs from the perspective of an S3 bucket which receives multiple
cloudtrail feeds from different accounts. This script discovers which
accounts are sending CloudTrails and then looks for when they last
supplied logs into the bucket.

Sends notifications to an SNS topic if logs haven't been delivered in
the last 24hr.